# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Taxi Ride Matching API
  description: |
    REST API for creating users and ride requests, discovering route-compatible
    travelers, completing two-sided match acceptance, and messaging in the private
    chat created for an accepted match.

    Typical workflow:
    1. Create two users.
    2. Create one future ride for each user. Matching runs when a ride is created.
    3. Retrieve a ride's potential matches.
    4. Both participating users accept the same match.
    5. Use the returned chat ID to send and retrieve messages.

    Authentication is not implemented in this MVP. Acting user IDs are supplied in
    request bodies and must belong to the relevant ride or chat.
  contact:
    name: Taxi API maintainers
  license:
    name: Private project
    identifier: LicenseRef-Proprietary
  version: 1.0.0-MVP
...
servers:
- url: http://localhost:8080
  description: Local development
...
paths:
  /api/rides/{id}:
    get:
      tags:
      - Rides
      summary: Get a ride request
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Ride found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RideResponse"
        "404":
          description: Ride not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RideResponse"
    put:
      tags:
      - Rides
      summary: Update an open ride
      description: "Re-resolves locations and route data, cancels stale pending matches,\
        \ and reruns matching. Only OPEN rides can be updated."
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRideRequest"
        required: true
      responses:
        "200":
          description: Ride updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RideResponse"
        "400":
          description: Validation or location failure
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RideResponse"
        "404":
          description: Ride not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RideResponse"
        "409":
          description: Ride is not open
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RideResponse"
    delete:
      tags:
      - Rides
      summary: Cancel a ride
      description: Performs a soft cancellation by changing status to CANCELLED and
        cancelling pending matches; no database row is deleted.
      operationId: cancel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Ride cancelled
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RideResponse"
        "404":
          description: Ride not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RideResponse"
        "409":
          description: Ride cannot be cancelled in its current state
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RideResponse"
...
  /api/rides:
    post:
      tags:
      - Rides
      summary: Create a ride request
      description: "Geocodes pickup/destination, calculates a driving route, stores\
        \ the OPEN ride, and immediately evaluates compatible OPEN rides owned by\
        \ other users. departureTime must be a future ISO-8601 instant."
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRideRequest"
        required: true
      responses:
        "201":
          description: Ride created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RideResponse"
        "400":
          description: Invalid request or location
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RideResponse"
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RideResponse"
        "502":
          description: Map provider failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RideResponse"
        "503":
          description: Map provider unavailable or quota exceeded
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RideResponse"
  /api/matches/{matchId}/reject:
    post:
      tags:
      - Matches
      summary: Reject a potential match
      description: Marks a PENDING match REJECTED. Rejected matches cannot later be
        accepted.
      operationId: reject
      parameters:
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchActionRequest"
        required: true
      responses:
        "200":
          description: Match rejected
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchActionResponse"
        "400":
          description: Acting user is not a participant
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchActionResponse"
        "404":
          description: Match not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchActionResponse"
        "409":
          description: Match is already terminal
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchActionResponse"
  /api/matches/{matchId}/accept:
    post:
      tags:
      - Matches
      summary: Accept a potential match
      description: "Records acceptance for one participating user. The first acceptance\
        \ leaves the match PENDING. The second acceptance atomically marks it ACCEPTED,\
        \ marks both rides MATCHED, and creates exactly one private chat."
      operationId: accept
      parameters:
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchActionRequest"
        required: true
      responses:
        "200":
          description: Decision recorded; chatId is populated after both accept
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchActionResponse"
        "400":
          description: Acting user is not a participant
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchActionResponse"
        "404":
          description: Match not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchActionResponse"
        "409":
          description: Match or ride state prevents acceptance
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchActionResponse"
  /api/chats/{chatId}/messages:
    get:
      tags:
      - Chats
      summary: List chat messages
      description: Returns a pageable message history. Defaults to 50 messages ordered
        by sentAt ascending. Use page and size query parameters for pagination.
      operationId: getMessages
      parameters:
      - name: chatId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: Message page returned
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageMessageResponse"
        "404":
          description: Chat not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageMessageResponse"
    post:
      tags:
      - Chats
      summary: Send a chat message
      description: Creates a message when senderId belongs to this chat. Messages
        must contain 1–4000 non-blank characters.
      operationId: sendMessage
      parameters:
      - name: chatId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
        required: true
      responses:
        "201":
          description: Message sent
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation failed or sender is not a participant
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Chat or sender not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/users/{userId}/rides:
    get:
      tags:
      - Rides
      summary: List a user's rides
      description: Returns all ride requests owned by the supplied user.
      operationId: getForUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Ride list returned
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RideResponse"
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RideResponse"
  /api/users/{userId}/chats:
    get:
      tags:
      - Chats
      summary: List a user's chats
      description: Returns chats in which the supplied user is a participant.
      operationId: getForUser_1
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Chat list returned
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatResponse"
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatResponse"
...
  /api/chats/{chatId}:
    get:
      tags:
      - Chats
      summary: Get a chat
      description: Returns chat metadata and both participant UUIDs.
      operationId: get_2
      parameters:
      - name: chatId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Chat found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "404":
          description: Chat not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChatResponse"
components:
  schemas:
    UpdateRideRequest:
      type: object
      description: Replacement trip details for an existing OPEN ride
      properties:
        pickupAddress:
          type: string
          example: "Arlington Heights, IL"
          maxLength: 500
          minLength: 0
        destinationAddress:
          type: string
          example: "Millennium Park, Chicago, IL"
          maxLength: 500
          minLength: 0
        departureTime:
          type: string
          format: date-time
          example: 2030-08-17T15:50:28.760Z
      required:
      - departureTime
      - destinationAddress
      - pickupAddress
    RideResponse:
      type: object
      description: Stored ride including resolved coordinates and compact route data
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        pickupAddress:
          type: string
        pickupLatitude:
          type: number
        pickupLongitude:
          type: number
        pickupPlaceId:
          type: string
        destinationAddress:
          type: string
        destinationLatitude:
          type: number
        destinationLongitude:
          type: number
        destinationPlaceId:
          type: string
        routeDistanceMeters:
          type: integer
          format: int64
        routeDurationSeconds:
          type: integer
          format: int64
        routeEncodedPolyline:
          type: string
        departureTime:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - OPEN
          - MATCHED
          - CANCELLED
          - COMPLETED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateRideRequest:
      type: object
      description: A user's planned trip. Addresses are resolved by the configured
        map provider.
      properties:
        userId:
          type: string
          format: uuid
          example: 3d4692be-cb4d-439d-ba31-5088fa217cd6
        pickupAddress:
          type: string
          example: "Schaumburg, IL"
          maxLength: 500
          minLength: 0
        destinationAddress:
          type: string
          example: "Millennium Park, Chicago, IL"
          maxLength: 500
          minLength: 0
        departureTime:
          type: string
          format: date-time
          description: Future UTC instant
          example: 2030-08-17T15:40:28.760Z
      required:
      - departureTime
      - destinationAddress
      - pickupAddress
      - userId
    MatchActionRequest:
      type: object
      description: Identifies which matched user is accepting or rejecting
      properties:
        userId:
          type: string
          format: uuid
          example: 3d4692be-cb4d-439d-ba31-5088fa217cd6
      required:
      - userId
    MatchActionResponse:
      type: object
      description: Current match state after a decision; chatId is null until both
        users accept
      properties:
        matchId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - PENDING
          - ACCEPTED
          - REJECTED
          - CANCELLED
        chatId:
          type: string
          format: uuid
    SendMessageRequest:
      type: object
      description: Message submitted by a participant in an accepted-match chat
      properties:
        senderId:
          type: string
          format: uuid
          example: 3d4692be-cb4d-439d-ba31-5088fa217cd6
        message:
          type: string
          example: See you at the pickup point
          maxLength: 4000
          minLength: 0
      required:
      - message
      - senderId
    MessageResponse:
      type: object
      description: Persisted chat message
      properties:
        id:
          type: string
          format: uuid
        chatId:
          type: string
          format: uuid
        senderId:
          type: string
          format: uuid
        message:
          type: string
        sentAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
    ChatResponse:
      type: object
      description: Private chat created exactly once for an accepted ride match
      properties:
        id:
          type: string
          format: uuid
        rideMatchId:
          type: string
          format: uuid
        participantIds:
          type: array
          items:
            type: string
            format: uuid
        createdAt:
          type: string
          format: date-time
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageMessageResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        sort:
          $ref: "#/components/schemas/SortObject"
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/MessageResponse"
        number:
          type: integer
          format: int32
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean